WebChatbot/requirements.txt
gradio==4.40.0
openai==1.39.0
python-dotenv==1.0.1
